finished js but needs adjusting
Finished writing the javascript, however its not fully functional currently. this is because Vue seems to be very particular about curly braces, unlike regular Javascript. In Vue javascript code, when defining things like "data" or "methods" or functions, the beginning curly brace needs to be on the same line as the declaration of that thing; the beginning curly brace cannot be on the next line. For example, it can't look like this:

data: function()
{
	return
	{
		disabled: false
	};
},

Instead, it has to look like this:

data: function() {
    return {
        disabled: false
    };
},

Notice the difference in terms of curly braces. to complete this project, I'll need to go back and adjust all the curly braces accordingly.